requirements-dev.txt
-r requirements.txt

# Supported libs
google-cloud-datastore>=1.0.0,<2
google-cloud-storage>=1.1.1,<2

# Testing
futures
httmock
pytest>=3
pytest-cov
tox

# Misc
bumpversion
